+++ tests/contrib/botocore/test_bedrock.py
@@ -14,78 +14,9 @@ from tests.subprocesstest import SubprocessTestCase
 from tests.subprocesstest import run_in_subprocess
 from tests.utils import DummyTracer
 from tests.utils import DummyWriter
-from tests.utils import override_config
-from tests.utils import override_global_config
 
 
 vcr = pytest.importorskip("vcr")
-
-
-@pytest.fixture(scope="session")
-def request_vcr():
-    yield get_request_vcr()
-
-
-@pytest.fixture
-def ddtrace_global_config():
-    config = {}
-    return config
-
-
-def default_global_config():
-    return {"_dd_api_key": "<not-a-real-api_key>"}
-
-
-@pytest.fixture
-def ddtrace_config_botocore():
-    return {}
-
-
-@pytest.fixture
-def aws_credentials():
-    """Mocked AWS Credentials. To regenerate test cassettes, comment this out and use real credentials."""
-    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
-    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
-    os.environ["AWS_SECURITY_TOKEN"] = "testing"
-    os.environ["AWS_SESSION_TOKEN"] = "testing"
-    os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
-
-
-@pytest.fixture
-def boto3(aws_credentials, mock_llmobs_span_writer, ddtrace_global_config, ddtrace_config_botocore):
-    global_config = default_global_config()
-    global_config.update(ddtrace_global_config)
-    with override_global_config(global_config):
-        with override_config("botocore", ddtrace_config_botocore):
-            patch()
-            import boto3
-
-            yield boto3
-            unpatch()
-
-
-@pytest.fixture
-def bedrock_client(boto3, request_vcr):
-    session = boto3.Session(
-        aws_access_key_id=os.getenv("AWS_ACCESS_KEY_ID", ""),
-        aws_secret_access_key=os.getenv("AWS_SECRET_ACCESS_KEY", ""),
-        aws_session_token=os.getenv("AWS_SESSION_TOKEN", ""),
-        region_name=os.getenv("AWS_DEFAULT_REGION", "us-east-1"),
-    )
-    bedrock_client = session.client("bedrock-runtime")
-    yield bedrock_client
-
-
-@pytest.fixture
-def mock_llmobs_span_writer():
-    patcher = mock.patch("ddtrace.llmobs._llmobs.LLMObsSpanWriter")
-    try:
-        LLMObsSpanWriterMock = patcher.start()
-        m = mock.MagicMock()
-        LLMObsSpanWriterMock.return_value = m
-        yield m
-    finally:
-        patcher.stop()
 
 
 class TestBedrockConfig(SubprocessTestCase):
@@ -111,6 +42,10 @@ class TestBedrockConfig(SubprocessTestCase):
         pin.override(self.bedrock_client, tracer=self.mock_tracer)
 
         super(TestBedrockConfig, self).setUp()
+
+    def tearDown(self):
+        unpatch()
+        super(TestBedrockConfig, self).tearDown()
 
     @run_in_subprocess(
         env_overrides=dict(
